now npc menu is in main game
