+++ source/core/src/main/com/deco2800/game/components/maingame/MainGameExitDisplay.java
@@ -1,10 +1,15 @@
 package com.deco2800.game.components.maingame;
 
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.scenes.scene2d.Actor;
+import com.badlogic.gdx.scenes.scene2d.ui.Button;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
+import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
 import com.deco2800.game.ui.UIComponent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -16,7 +21,8 @@ public class MainGameExitDisplay extends UIComponent {
   private static final Logger logger = LoggerFactory.getLogger(MainGameExitDisplay.class);
   private static final float Z_INDEX = 2f;
   private Table table;
-
+  private static final int NPC_MENU_BUTTON_WIDTH = 150;
+  private static final int NPC_MENU_BUTTON_HEIGHT = 80;
   @Override
   public void create() {
     super.create();
@@ -27,7 +33,14 @@ public class MainGameExitDisplay extends UIComponent {
     table = new Table();
     table.top().right();
     table.setFillParent(true);
-
+    /** build new style eviction menu button */
+    Button.ButtonStyle styleEvictionMenu = new Button.ButtonStyle();
+    styleEvictionMenu.over = new TextureRegionDrawable(new TextureRegion(
+            new Texture(Gdx.files.internal("images/eviction_menu/menuIcon_black.png"))));
+    //here is for button effect when you pressed on button
+    styleEvictionMenu.up = new TextureRegionDrawable(new TextureRegion(
+            new Texture(Gdx.files.internal("images/eviction_menu/menuIcon_white.png"))));
+    Button npcMenuBtn = new Button(styleEvictionMenu);
     TextButton mainMenuBtn = new TextButton("Exit", skin);
 
     // Triggers an event when the button is pressed.
@@ -39,8 +52,19 @@ public class MainGameExitDisplay extends UIComponent {
           entity.getEvents().trigger("exit");
         }
       });
+    npcMenuBtn.addListener(
+            new ChangeListener() {
+              @Override
+              public void changed(ChangeEvent changeEvent, Actor actor) {
+                logger.debug("Load button clicked");
+                entity.getEvents().trigger("NpcMenu");
+              }
+            });
 
     table.add(mainMenuBtn).padTop(10f).padRight(10f);
+    table.row();
+    table.add(npcMenuBtn).padTop(15f).width(NPC_MENU_BUTTON_WIDTH).height(NPC_MENU_BUTTON_HEIGHT);
+    table.row();
 
     stage.addActor(table);
   }
