+++ source/core/src/main/com/deco2800/game/components/mainmenu/MainMenuActions.java
@@ -20,7 +20,7 @@ public class MainMenuActions extends Component {
   @Override
   public void create() {
     entity.getEvents().addListener("start", this::onStart);
-    entity.getEvents().addListener("NpcMenu", this::onNpcMenu);
+
     entity.getEvents().addListener("exit", this::onExit);
     entity.getEvents().addListener("settings", this::onSettings);
   }
@@ -34,14 +34,6 @@ public class MainMenuActions extends Component {
 
   }
 
-  /**
-   * Intended for loading a saved game state.
-   * Load functionality is not actually implemented.
-   */
-  private void onNpcMenu() {
-    logger.info("Load game");
-    game.setScreen(GdxGame.ScreenType.NPC_EVICTION_MENU);
-  }
 
   /**
    * Exits the game.
