+++ source/core/src/main/com/deco2800/game/components/npcEvictionMenu/NpcEvictionMenuDisplay.java
@@ -170,7 +170,7 @@ public class NpcEvictionMenuDisplay extends UIComponent {
     }
 
     private void exitMenu() {
-        game.setScreen(GdxGame.ScreenType.MAIN_MENU);
+        game.setScreen(GdxGame.ScreenType.MAIN_GAME);
     }
 
     @Override
